+++ tests/pgnlibrary.rs
@@ -1,28 +1,13 @@
-
 extern crate canparse;
 
-use canparse::dbc::Entry;
-use canparse::pgn::{PgnLibrary, SpnDefinition, ParseMessage};
-use std::str::FromStr;
-use std::collections::HashMap;
+use canparse::pgn::{ParseMessage, PgnLibrary, SpnDefinition};
 
 #[test]
 fn pgnlib_build_parse() {
-    let mut lib = PgnLibrary::new( HashMap::default() );
-
-    let data: String = include_bytes!("./data/sample.dbc")
-        .iter().map(|b| *b as char).collect();
-
-    // Parse db lines into PgnLibrary
-    for line in data.lines() {
-        if let Ok(entry) = Entry::from_str(&line) {
-            lib.add_entry(entry).ok();
-        }
-    }
+    let lib = PgnLibrary::from_dbc_file("./tests/data/sample.dbc").unwrap();
 
     // Pull signal definition for engine speed
-    let enginespeed_def: &SpnDefinition = lib
-        .get_spn("Engine_Speed").unwrap();
+    let enginespeed_def: &SpnDefinition = lib.get_spn("Engine_Speed").unwrap();
 
     // Parse frame containing engine speed
     let msg: [u8; 8] = [0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88];
@@ -33,10 +18,12 @@ fn pgnlib_build_parse() {
 
 #[test]
 fn pgnlib_from_dbc_file() {
-
     let lib = PgnLibrary::from_dbc_file("./tests/data/sample.dbc");
     assert!(lib.is_ok(), "PgnLibrary should have built successfully.");
 
     let lib_fail = PgnLibrary::from_dbc_file("./tests/data/sample.dbc.fail");
-    assert_eq!(lib_fail.map_err(|e| e.kind()), Err(std::io::ErrorKind::NotFound))
+    assert_eq!(
+        lib_fail.map_err(|e| e.kind()),
+        Err(std::io::ErrorKind::NotFound)
+    )
 }